requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.60.0
pycocotools>=2.0.4
pytesseract>=0.3.10
easydict>=1.9
python-dotenv>=1.0.0